requirements.txt
ttkbootstrap>=1.0.0
plyer>=2.1.0
pandas>=2.0.0
matplotlib>=3.7.0
pytesseract>=0.3.10
Pillow>=9.5.0